Phone: Settings: do not override network mode on read
Users may use the network tile to put the network in modes not supported
by this checkbox (e.g. 3G). A side effect of this is that the tile will
display the wrong mode (2G3G instead of 3G) after entering the
Mobile network settings page because the checkbox writes 2G3G into the
system settings without actually changing it at the RIL.

Change-Id: I4658468caadb4782520856ef70feee7040065523